requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
dnspython==2.7.0
fastapi==0.115.12
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
jiter==0.9.0
motor==3.7.0
openai==1.72.0
pydantic==2.11.3
pydantic-core==2.33.1
pymongo==4.12.0
python-dotenv==1.1.0
requests==2.32.3
sniffio==1.3.1
starlette==0.46.1
tqdm==4.67.1
typing-extensions==4.13.2
typing-inspection==0.4.0
urllib3==2.4.0
uvicorn==0.34.0
